refactor: remove unused production helpers
