--- internal/adapter/config/testdata/render_operator_bootstrap.hcl
@@ -1,37 +1,0 @@
-initialize "operator-bootstrap" {
-  request "enable-jwt-auth" {
-    operation = "update"
-    path      = "sys/auth/jwt-operator"
-    data {
-      type        = "jwt"
-      description = "Auth method for OpenBao Operator"
-    }
-  }
-  request "config-jwt-auth" {
-    operation = "update"
-    path      = "auth/jwt-operator/config"
-    data {
-      bound_issuer           = "https://kubernetes.default.svc"
-      jwt_validation_pubkeys = ["-----BEGIN PUBLIC KEY-----\ntest-public-key\n-----END PUBLIC KEY-----\n"]
-    }
-  }
-  request "create-operator-policy" {
-    operation = "update"
-    path      = "sys/policies/acl/openbao-operator"
-    data {
-      policy = "path \"sys/health\" { capabilities = [\"read\"] }\npath \"sys/step-down\" { capabilities = [\"sudo\", \"update\"] }\npath \"sys/storage/raft/configuration\" { capabilities = [\"read\"] }\npath \"sys/storage/raft/remove-peer\" { capabilities = [\"update\"] }\npath \"sys/storage/raft/autopilot/configuration\" { capabilities = [\"read\", \"update\"] }\npath \"sys/storage/raft/autopilot/state\" { capabilities = [\"read\"] }"
-    }
-  }
-  request "create-operator-role" {
-    operation = "update"
-    path      = "auth/jwt-operator/role/openbao-operator"
-    data {
-      role_type       = "jwt"
-      user_claim      = "sub"
-      bound_audiences = ["openbao-internal"]
-      bound_subject   = "system:serviceaccount:openbao-operator-system:openbao-operator-controller"
-      token_policies  = ["openbao-operator"]
-      ttl             = "1h"
-    }
-  }
-}
